Remove TODO related to storing RSS feed link in queued job
The idea behind this TODO was, that when the bot is running feed links won't change, so they could be stored in queued job data, instead of parsed every time.
However, it would complicate logic behind the job and require new function in feed_reader module.
Also, I quite like the fact, that any RSS links are hidden behind the feed_reader module.