Read the latest feed entry ID from the DB when checking for updates
Read the latest feed entry ID from the DB when checking for feed updates, rather than storing it in the job data.
This way there's a smaller chance of some kind of de-synchronization between running jobs and the DB, the only data stored in the job is not changing.
There's also no need of updating job data after an update.