Assume that feeds without entries are not valid
When checking if a given feed is valid check that it does have any entries already, in addition to verifying response codes.
This effectively disallows subscribing for new feeds, which doesn't have any entries in them yet.
However, it allows for better detection of cases where feed became not valid.