Split commands using CallbackQueryHandler into two handlers
Each command using CallbackQueryHandler is split into two handlers:
 1. Initial command handler, which creates a message with inline buttons
 2. Followup conversation handler, which handles query from inline buttons
For "remove" and "removeall" commands this allows for entire conversation handler (and handling query in general) to be limited per message.
For "add" command the followup handler doesn't modify the query, only triggers reqular message command.
For all of them it results in limited odds of one query interacting with a different query of the same type.